+++ src/io/mod.rs
@@ -1,4 +1,3 @@
-pub mod initial_read;
 pub mod presenter;
 pub mod reader;
 pub mod routing;